// src/types/stylesTypes.ts
export type BoxDirections = {
  /**
   * Padding or margin on the top of an element in px.
   * @default 0
   */
  top: number;
  /**
   * Padding or margin on the right of an element in px.
   * @default 0
   */
  right: number;
  /**
   * Padding or margin on the bottom of an element in px.
   * @default 0
   */
  bottom: number;
  /**
   * Padding or margin on the left of an element in px.
   * @default 0
   */
  left: number;
};